updated the price validation

+++ server/models/products.js
@@ -5,7 +5,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     price: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(7, 2),
       allowNull: false,
     },
     currency: {
